prov/efa: move completion reporting to rxr_op_entry.c
This patch moved the following functions:

   rxr_cq_write_rx_error()
   rxr_cq_write_tx_error()
   rxr_cq_write_rx_completion()
   rxr_cq_write_tx_completion()
   rxr_cq_handle_send_completion()
   rxr_cq_complete_recv()

from rxr_cq.c to rxr_op_entry.c, and renamed them to

   rxr_rx_entry_handle_error()
   rxr_tx_entry_handle_error()
   rxr_rx_entry_report_completion()
   rxr_tx_entry_report_completion()
   rxr_op_entry_handle_send_completed()
   rxr_op_entry_handle_recv_completed()

The renaming is because these functions operates on
rxr_op_entry instead of on rxr_cq. The function header
comments have been updated.

This patch also introduced a data member "ep" to
rxr_op_entry and initialized "ep" to be pointing
to rxr_ep. The "ep" argument was then removed from
arguments of the 6 above mentioned functions.
